requirements.txt
discord.py>=2.4.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
tortoise-orm>=0.20.0  # Async ORM
aiosqlite>=0.17.0    # SQLite driver
asyncpg>=0.29.0      # PostgreSQL driver
pydantic>=2.5.0      # Data validation
Pillow>=10.0.0
